Change Sexp to hold Symbol pointer, reorg code, added tests.
The Sexp struct was changed so that it now holds a pointer to a Symbol struct
rather than a value of a Symbol struct.

The code was reorganized and split into multiple files.

"test-sexp.c" was added which contains some basic tests for S-expression
parsing.

The makefile was updated to reflect the above changes.